[rpc] Add unix time at boot as value for node metadata rpc - need direct machine uptime
