+++ internal/hmyapi/apiv1/harmony.go
@@ -6,9 +6,7 @@ import (
 
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/harmony-one/harmony/api/proto"
-	nodeconfig "github.com/harmony-one/harmony/internal/configs/node"
-	"github.com/harmony-one/harmony/internal/params"
-	"github.com/harmony-one/harmony/shard"
+	commonRPC "github.com/harmony-one/harmony/internal/hmyapi/common"
 )
 
 // PublicHarmonyAPI provides an API to access Harmony related information.
@@ -45,51 +43,7 @@ func (s *PublicHarmonyAPI) GasPrice(ctx context.Context) (*hexutil.Big, error) {
 	return (*hexutil.Big)(big.NewInt(1)), nil
 }
 
-// NodeMetadata captures select metadata of the RPC answering node
-type NodeMetadata struct {
-	BLSPublicKey   []string           `json:"blskey"`
-	Version        string             `json:"version"`
-	NetworkType    string             `json:"network"`
-	ChainConfig    params.ChainConfig `json:"chain-config"`
-	IsLeader       bool               `json:"is-leader"`
-	ShardID        uint32             `json:"shard-id"`
-	CurrentEpoch   uint64             `json:"current-epoch"`
-	BlocksPerEpoch *uint64            `json:"blocks-per-epoch,omitempty"`
-	Role           string             `json:"role"`
-	DNSZone        string             `json:"dns-zone"`
-	Archival       bool               `json:"is-archival"`
-}
-
 // GetNodeMetadata produces a NodeMetadata record, data is from the answering RPC node
-func (s *PublicHarmonyAPI) GetNodeMetadata() NodeMetadata {
-	cfg := nodeconfig.GetDefaultConfig()
-	header := s.b.CurrentBlock().Header()
-	var blockEpoch *uint64
-
-	if header.ShardID() == shard.BeaconChainShardID {
-		sched := shard.Schedule.InstanceForEpoch(header.Epoch())
-		b := sched.BlocksPerEpoch()
-		blockEpoch = &b
-	}
-
-	blsKeys := []string{}
-	if cfg.ConsensusPubKey != nil {
-		for _, key := range cfg.ConsensusPubKey.PublicKey {
-			blsKeys = append(blsKeys, key.SerializeToHexStr())
-		}
-	}
-
-	return NodeMetadata{
-		blsKeys,
-		nodeconfig.GetVersion(),
-		string(cfg.GetNetworkType()),
-		*s.b.ChainConfig(),
-		s.b.IsLeader(),
-		s.b.GetShardID(),
-		header.Epoch().Uint64(),
-		blockEpoch,
-		cfg.Role().String(),
-		cfg.DNSZone,
-		cfg.GetArchival(),
-	}
+func (s *PublicHarmonyAPI) GetNodeMetadata() commonRPC.NodeMetadata {
+	return s.b.GetNodeMetadata()
 }
